+++ 1.Domain/Domain.Services.Impl/Domain.Services.Impl.csproj
@@ -20,6 +20,7 @@
     <PackageReference Include="Google.Apis.Calendar.v3" Version="1.40.2.1643" />
     <PackageReference Include="Google.Apis.Core" Version="1.43.0" />
     <PackageReference Include="Google.Apis.Drive.v3" Version="1.43.0.1863" />
+    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.2.3" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.2.0" />
