+++ 1.Domain/Domain.Services/Interfaces/Services/INotificationService.cs
@@ -1,0 +1,12 @@
+﻿using Domain.Model;
+using System.Collections.Generic;
+
+namespace Domain.Services.Interfaces.Services
+{
+    public interface INotificationService
+    {
+       List<Notification> GetNotification();
+
+       void ReadNotification(int id);
+    }
+}
