+++ 2.Infrastructure/Domain.Services.Repositories.EF/NotificationRepository.cs
@@ -19,7 +19,6 @@ namespace Domain.Services.Repositories.EF
 
         public void Create(Notification notification, int candId)
         {
-            //TODO: Assign notification to users 
             var candList = GetReferralsList(candId);
             foreach (var cand in candList)
             {
